// src/types.ts
export interface OrganicRentalEdge {
  __typename: string;
  node: SearchRentalListing;
  amenitiesMatch: boolean;
  matchedAmenities: string[];
  missingAmenities: string[];
}

export interface FeaturedRentalEdge {
  __typename: string;
  node: SearchRentalListing;
  amenitiesMatch: boolean;
  matchedAmenities: string[];
  missingAmenities: string[];
}

export interface SponsoredRentalEdge {
  __typename: string;
  node: SearchRentalListing;
  sponsoredSimilarityLabel: string;
}

export type RentalEdge =
  | OrganicRentalEdge
  | FeaturedRentalEdge
  | SponsoredRentalEdge;

export interface SearchRentalsResponse {
  searchRentals: {
    __typename: string;
    edges: RentalEdge[];
    totalCount: number;
  };
}

export interface NumberRange {
  lowerBound: number | null;
  upperBound: number | null;
}

export interface SearchFilters {
  areas?: AreaCode[];
  rentalStatus?: "ACTIVE";
  price?: NumberRange;
  bedrooms?: NumberRange;
  bathrooms?: NumberRange;
  amenities?: Amenity[];
  petsAllowed?: boolean;
  noFee?: boolean;
}

export interface Sorting {
  attribute: "RECOMMENDED" | "PRICE" | "LISTED_AT" | "INTERESTING_CHANGE_AT" | "SQFT";
  direction: "ASCENDING" | "DESCENDING";
}

export interface SearchRentalsInput {
  sorting?: Sorting;
  adStrategy?: "NONE";
  filters: SearchFilters;
  perPage?: number;
  page?: number;
  userSearchToken?: string;
}